Cleaned code and fixes
