+++ addons/de/xrechnung/invoices.go
@@ -4,7 +4,6 @@ import (
 	"github.com/invopop/gobl/bill"
 	"github.com/invopop/gobl/cbc"
 	"github.com/invopop/gobl/org"
-	"github.com/invopop/gobl/tax"
 	"github.com/invopop/validation"
 )
 
@@ -158,28 +157,6 @@ func validateDeliveryAddress(value interface{}) error {
 			validation.Required,
 		),
 	)
-}
-
-func validateTaxCombo(tc *tax.Combo) error {
-	if tc == nil {
-		return nil
-	}
-	return validation.ValidateStruct(tc,
-		validation.Field(&tc.Category,
-			validation.When(tc.Category == tax.CategoryVAT,
-				validation.By(validateVATRate),
-			),
-		),
-	)
-}
-
-// BR-DE-14
-func validateVATRate(value interface{}) error {
-	rate, _ := value.(cbc.Key)
-	if rate == "" {
-		return validation.NewError("required", "VAT category rate is required")
-	}
-	return nil
 }
 
 func validateCorrectiveInvoice(value interface{}) error {
